requirements.txt
flask
requests
bs4
httpx[http2]
python-dotenv
gunicorn
